src/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=3.10,<4.0
pytest>=6.2,<7.0
pytest-django>=4.5,<5.0
requests>=2.25,<3.0